fix golang ci-lint
